// katchimeras/constants/creature-hatchling-alpha-bounds.gen.ts
// Generated by scripts/generate-hatchling-lod.py. Do not hand edit.
import type { HomeVisualKey } from '@/types/home';

export type CreatureHatchlingAlphaBounds = { left: number; top: number; right: number; bottom: number };

export const CREATURE_HATCHLING_ALPHA_THRESHOLD = 16;
export const CREATURE_HATCHLING_ALPHA_BOUNDS: Partial<Record<HomeVisualKey, CreatureHatchlingAlphaBounds>> = {
  bedrotte: { left: 0.210938, top: 0.072266, right: 0.793945, bottom: 0.948242 },
  cheerlet: { left: 0.210938, top: 0.064453, right: 0.794922, bottom: 0.900391 },
  feastle: { left: 0.177734, top: 0.140625, right: 0.829102, bottom: 0.894531 },
  flickerbun: { left: 0.261719, top: 0.116211, right: 0.737305, bottom: 0.872070 },
  gatherglow: { left: 0.271484, top: 0.100586, right: 0.736328, bottom: 0.888672 },
  mossprout: { left: 0.135742, top: 0.047852, right: 0.864258, bottom: 0.953125 },
  pagelet: { left: 0.134766, top: 0.088867, right: 0.857422, bottom: 0.867188 },
  relicoon: { left: 0.237305, top: 0.094727, right: 0.760742, bottom: 0.898438 },
  shellio: { left: 0.195312, top: 0.098633, right: 0.804688, bottom: 0.888672 },
  skylo: { left: 0.235352, top: 0.097656, right: 0.760742, bottom: 0.897461 },
  steppling: { left: 0.261719, top: 0.084961, right: 0.740234, bottom: 0.886719 },
  tasklet: { left: 0.271484, top: 0.046875, right: 0.776367, bottom: 0.932617 },
  vesperitt: { left: 0.240234, top: 0.096680, right: 0.749023, bottom: 0.892578 },
};